This commit contains alot of modifications and alot of features has been added to the app, it's almost finished, just some final touches and it will be ver cool and ready to use, It was a really really funny journey and I learned alot from it

+++ app/src/main/java/com/example/quizapp/model/QuizListModel.java
@@ -6,13 +6,13 @@ public class QuizListModel {
 
     @DocumentId
     private String quiz_id;
-    private long questions;
+    private int questions;
     private String name , image, desc, level, visibility;
 
     public QuizListModel() {
     }
 
-    public QuizListModel(String quiz_id, long questions, String name, String image, String desc, String level, String visibility) {
+    public QuizListModel(String quiz_id, int questions, String name, String image, String desc, String level, String visibility) {
         this.quiz_id = quiz_id;
         this.questions = questions;
         this.name = name;
@@ -34,7 +34,7 @@ public class QuizListModel {
         return questions;
     }
 
-    public void setQuestions(long questions) {
+    public void setQuestions(int questions) {
         this.questions = questions;
     }
